Test all plugins when yarn.lock file changes
Since the plugins, app, and backend share a lockfile, it seems we can
get some odd issues when we update dependencies that are not caught
until the next time a plugin is tested.

In particular, this happened with Dependabot, and although we were
updating the package.json as well to make sure affected plugins were
tested I think the issue comes with updating dependencies of
dependencies not directly listed in the package.json.